Add files via upload
This code imports the required libraries and loads customer data from a CSV file. It then splits the data into training and test sets, normalizes the training data, and performs K-means clustering with different values of k to find the optimal number of clusters. The code then applies K-means clustering with the best value of k to the test data and visualizes the clustering results using a scatter plot with age and income as the two features.